+++ changelog.md
@@ -1,3 +1,15 @@
+# BetterBlizzFrames 1.5.6b
+## The War Within
+### New stuff:
+- Hide Resource/Power setting now has individual class settings. Access them by rightclicking the checkbox.
+- The "Normal Evoker Empowered Castbar" setting now also makes sArena castbars normal.
+- Misc: Hide Talking Heads setting.
+### Tweak:
+- Nahj profile update
+### Bugfix:
+- Fix an issue with the red threat border overlapping auras with ClassicFrames addon.
+- Fix issue with the new Misc settings for Queue Status Eye from last patch making the eye unclickable with the settings on.
+
 # BetterBlizzFrames 1.5.6
 ## The War Within
 ### New stuff: